+++ packages/salesforcedx-sobjects-faux-generator/src/describe/types.ts
