+++ packages/salesforcedx-sobjects-faux-generator/src/types/general.ts
@@ -4,7 +4,7 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import { SObjectShortDescription } from '../describe';
+import { SObjectShortDescription } from '../describe/types';
 import { SObject } from './describe';
 
 export type SObjectCategory = 'ALL' | 'STANDARD' | 'CUSTOM';
